Avoid using cpio-filter and cpio --append
We seem to go out of our way to avoid creating a 'new' cpio archive.
This all seems rather unnecessary and error prone and cpio-filter
will create a new archive anyway, so it's also rather redundant.

Adopt a simpler process: extract the initrd fully, and run the plymouth
population in-place, then just recompress the result.

This also changes the archive creation stage to be the same commands
used by dracut for clarity.

git-svn-id: svn://svn.mageia.org/svn/soft/bootsplash/trunk@4047 5b7c3c08-40e5-403b-9995-ace06908e4af