Add support for xz compressed initrds.
Default options for xz compression taken from dracut.

This change also removes various /bin/ path prefixes as there was
previously a mix. xz* tools are technically in /usr so thus support
may not be 100% complete until we have finished the /usr merge in
mga3 and as such gzip compressed initrds should likely remain the default
for mga2.

No support for xz compressed initrds has yet been added to the remove-boot-splash
script which is inexplicably implemented in perl rather than in bash
like the rest of the scripts.

git-svn-id: svn://svn.mageia.org/svn/soft/bootsplash/trunk@4049 5b7c3c08-40e5-403b-9995-ace06908e4af