Do not pass theme in when calling make-boot-splash-raw.
This reverts and earlier commit that used it explicitly, however this
approach is more error prone.

The default theme is set in make-boot-splash, and thus it will be used
when calling via this route. Error checking is done at this higher level
and thus if an invalid theme is passed in the old theme will still be default
and we will properly regenerate initrd.

Clarify that the theme argument should not be passed in the help output.

git-svn-id: svn://svn.mageia.org/svn/soft/bootsplash/trunk@4051 5b7c3c08-40e5-403b-9995-ace06908e4af