Do not remove dracut hooks.
Most hooks are harmless and do checks. One hook will fail
if /bin/plymouthd is removed but I will fix that at the dracut
end as the hook is not replaced when re-populating the initrd again
so we should just make the hook a noop when the binary does not
exist instead.

git-svn-id: svn://svn.mageia.org/svn/soft/bootsplash/trunk@4052 5b7c3c08-40e5-403b-9995-ace06908e4af